(bug)fix temporal profile time delta calculation (#866)
* fix temporal profile time delta calculation

* max intensity bug

+++ controller/pwm_profile/random.go
@@ -16,7 +16,7 @@ type random struct {
 	temporal
 	previous     float64
 	peaks        []float64
-	peakInterval int
+	peakInterval float64
 }
 
 func (r *random) Name() string {
@@ -33,7 +33,7 @@ func Random(conf json.RawMessage, min, max float64) (*random, error) {
 
 func NewRandom(t temporal) *random {
 	rand.Seed(seed)
-	numPeaks := t.TotalSeconds() / peakInterval
+	numPeaks := int(t.TotalSeconds() / peakInterval)
 	if numPeaks == 0 {
 		numPeaks = 1
 	}
@@ -53,13 +53,13 @@ func (s *random) Get(t time.Time) float64 {
 	if s.IsOutside(t) {
 		return 0
 	}
-	i := s.PastSeconds(t) / s.peakInterval
+	i := int(s.PastSeconds(t) / s.peakInterval)
 	prevPeak := s.peaks[i]
 	nextPeak := s.peaks[0]
 	if (i + 1) < len(s.peaks) {
 		nextPeak = s.peaks[i+1]
 	}
-	f := (nextPeak - prevPeak) / float64(s.peakInterval)
+	f := (nextPeak - prevPeak) / s.peakInterval
 	inc := rand.NormFloat64() + f
 	s.previous += inc
 	if s.previous > s.max {
